requirements.txt
streamlit>=1.20
openai-whisper==1.1.10
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.2
yake>=0.4.8
pydub>=0.25.1
numpy>=1.23
pandas>=1.5
tqdm>=4.65
torch>=2.0
faiss-cpu>=1.7.4 # optional for faster similarity search; remove if install
issues
python-multipart>=0.0.5
